test(best-practices): added `buildTest()` utility function and migrated BP1 to use it
